+++ src/options/common.rs
@@ -3,6 +3,7 @@ mod cmd;
 mod logging;
 mod network;
 mod search;
+mod shell;
 mod time;
 mod value;
 
@@ -11,5 +12,6 @@ pub use cmd::*;
 pub use logging::*;
 pub use network::*;
 pub use search::*;
+pub use shell::*;
 pub use time::*;
 pub use value::*;